refactor: extract Weblurk widget to dedicated file
